utils: document ENOBUFS in nfqnl_test.c
This patch documents the ENOBUFS error in the example file, that
is a common problem is that question over and over again in the
mailing list.

I (Pablo) have mangled this patch with some comestic cleanups. BTW,
Mistick Levi sent a similar patch in the same timeline (amazing how
sometimes the same works can clash).

Signed-off-by: Alessandro Vesely <vesely@tana.it>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>